+++ packages/server/src/unstable-core-do-not-import/clientish/inference.ts
@@ -1,6 +1,11 @@
 import type { inferObservableValue } from '@trpc/server/observable';
 import type { AnyProcedure } from '../procedure';
-import type { ClientInferrable, inferRootTypes } from './inferrable';
+import type { AnyRouter, RouterRecord } from '../router';
+import type {
+  AnyClientRootTypes,
+  ClientInferrable,
+  inferRootTypes,
+} from './inferrable';
 import type { Serialize } from './serialize';
 
 /**
@@ -32,3 +37,31 @@ export type inferProcedureParams<TProcedure> = TProcedure extends AnyProcedure
   : never;
 export type inferProcedureOutput<TProcedure> =
   inferProcedureParams<TProcedure>['_output_out'];
+
+export type GetInferenceHelpers<
+  TType extends 'input' | 'output',
+  TRoot extends AnyClientRootTypes,
+  TRecord extends RouterRecord,
+> = {
+  [TKey in keyof TRecord]: TRecord[TKey] extends infer $Value
+    ? $Value extends RouterRecord
+      ? GetInferenceHelpers<TType, TRoot, $Value>
+      : $Value extends AnyProcedure
+      ? TType extends 'input'
+        ? inferProcedureInput<$Value>
+        : inferTransformedProcedureOutput<TRoot, $Value>
+      : never
+    : never;
+};
+
+export type inferRouterInputs<TRouter extends AnyRouter> = GetInferenceHelpers<
+  'input',
+  TRouter['_def']['_config']['$types'],
+  TRouter['_def']['record']
+>;
+
+export type inferRouterOutputs<TRouter extends AnyRouter> = GetInferenceHelpers<
+  'output',
+  TRouter['_def']['_config']['$types'],
+  TRouter['_def']['record']
+>;
