+++ packages/server/src/unstable-core-do-not-import/index.ts
@@ -52,9 +52,7 @@ export type {
   createRouterFactory,
   inferRouterContext,
   inferRouterError,
-  inferRouterInputs,
   inferRouterMeta,
-  inferRouterOutputs,
 } from './router';
 
 export * from './clientish/inference';
